requirements.txt
appdirs==1.4.3
astroid==2.4.2
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
black==19.10b0
blinker==1.4
Brotli==1.0.7
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cycler==0.10.0
dash==1.14.0
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
distlib==0.3.0
Django==2.1.5
filelock==3.0.12
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Compress==1.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
future==0.18.2
idna==2.8
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
numpy==1.18.1
pandas==1.1.0
pandas-datareader==0.9.0
pathspec==0.8.0
Pillow==7.0.0
plotly==4.9.0
pycparser==2.19
pylint==2.5.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
regex==2020.7.14
requests==2.22.0
retrying==1.3.3
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==20.0.4
Werkzeug==1.0.0
wrapt==1.12.1
WTForms==2.2.1
zipp==2.2.0